Moved server code into its own file.
